Refactor Configs into appropriate packages
