+++ src/test/java/at/hannibal2/skyhanni/test/garden/VisitorToolTipParserTest.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.test.garden
 
-import at.hannibal2.skyhanni.config.features.GardenConfig
+import at.hannibal2.skyhanni.config.features.garden.GardenConfig
 import at.hannibal2.skyhanni.features.garden.visitor.VisitorTooltipParser
 import org.junit.jupiter.api.Test
 
@@ -22,7 +22,9 @@ class VisitorToolTipParserTest {
 
     @Test
     fun testParseItemsNeeded() {
-        val parsedData = VisitorTooltipParser.parse(lore, GardenConfig())
+        val parsedData = VisitorTooltipParser.parse(lore,
+            GardenConfig()
+        )
         assert(parsedData.itemsNeeded.isNotEmpty()) {
             "Visitor items needed is ${parsedData.itemsNeeded.count()} instead of 1"
         }
@@ -33,7 +35,9 @@ class VisitorToolTipParserTest {
 
     @Test
     fun testParseRewards() {
-        val parsedData = VisitorTooltipParser.parse(lore, GardenConfig())
+        val parsedData = VisitorTooltipParser.parse(lore,
+            GardenConfig()
+        )
         assert(parsedData.rewards.isNotEmpty()) {
             "Visitor rewards is ${parsedData.rewards.count()} instead of 6"
         }
@@ -56,7 +60,9 @@ class VisitorToolTipParserTest {
 
     @Test
     fun testParseCopper() {
-        val parsedData = VisitorTooltipParser.parse(lore, GardenConfig())
+        val parsedData = VisitorTooltipParser.parse(lore,
+            GardenConfig()
+        )
         val copper = parsedData.rewards.get("Copper")
         assert(copper == 23) {
             "Visitor rewards does not contain 'Copper' with amount '23'"
